+++ src/module/cellular.rs
@@ -3,6 +3,7 @@
 use calloop::LoopHandle;
 use calloop::channel::Event;
 
+use crate::config::ConfigPanelModule;
 use crate::dbus::modem_manager::{self, ModemConnection};
 use crate::module::{Alignment, DrawerModule, Module, PanelModule, PanelModuleContent, Toggle};
 use crate::text::Svg;
@@ -14,10 +15,12 @@ pub struct Cellular {
 
     /// Desired connectivity state.
     desired_enabled: bool,
+
+    alignment: Alignment,
 }
 
 impl Cellular {
-    pub fn new(event_loop: &LoopHandle<'static, State>) -> Result<Self> {
+    pub fn new(event_loop: &LoopHandle<'static, State>, alignment: Alignment) -> Result<Self> {
         // Subscribe to ModemManager DBus events.
         let rx = modem_manager::modem_listener();
         event_loop.insert_source(rx, move |event, _, state| {
@@ -27,10 +30,10 @@ impl Cellular {
             };
 
             // Ignore updates that change nothing.
-            let module = &mut state.modules.cellular;
-            if connection == module.connection {
-                return;
-            }
+            let module = match state.modules.cellular.as_mut() {
+                Some(module) if connection != module.connection => module,
+                _ => return,
+            };
 
             let old_enabled = module.desired_enabled;
             let old_svg = module.svg();
@@ -40,12 +43,12 @@ impl Cellular {
             module.connection = connection;
 
             // Request redraw only if SVG changed.
-            if old_svg != state.modules.cellular.svg() || old_enabled != connection.enabled {
+            if old_svg != module.svg() || old_enabled != connection.enabled {
                 state.unstall();
             }
         })?;
 
-        Ok(Self { connection: ModemConnection::default(), desired_enabled: false })
+        Ok(Self { alignment, connection: ModemConnection::default(), desired_enabled: false })
     }
 }
 
@@ -61,11 +64,15 @@ impl Module for Cellular {
 
 impl PanelModule for Cellular {
     fn alignment(&self) -> Alignment {
-        Alignment::Right
+        self.alignment
     }
 
     fn content(&self) -> PanelModuleContent {
         PanelModuleContent::Svg(self.svg())
+    }
+
+    fn config_variant(&self) -> ConfigPanelModule {
+        ConfigPanelModule::Cellular
     }
 }